move config of api and blocksync to their own package
