+++ test/apitest/coreservice_test.go
@@ -3,7 +3,7 @@
 // or fitness for purpose and, to the extent permitted by law, all liability for your use of the code is disclaimed.
 // This source code is governed by Apache License 2.0 that can be found in the LICENSE file.
 
-package api
+package apitest
 
 import (
 	"context"
@@ -18,6 +18,7 @@ import (
 
 	"github.com/iotexproject/iotex-core/action"
 	"github.com/iotexproject/iotex-core/actpool"
+	"github.com/iotexproject/iotex-core/api"
 	"github.com/iotexproject/iotex-core/api/logfilter"
 	"github.com/iotexproject/iotex-core/blockchain"
 	"github.com/iotexproject/iotex-core/blockchain/blockdao"
@@ -31,7 +32,7 @@ func TestLogsInRange(t *testing.T) {
 	defer cleanCallback()
 
 	t.Run("blocks with four logs", func(t *testing.T) {
-		testData := &filterObject{FromBlock: "1", ToBlock: "4"}
+		testData := &api.FilterObjectWrapper{FromBlock: "1", ToBlock: "4"}
 		filter, err := getTopicsAddress(testData.Address, testData.Topics)
 		require.NoError(err)
 		from, err := strconv.ParseUint(testData.FromBlock, 10, 64)
@@ -45,7 +46,7 @@ func TestLogsInRange(t *testing.T) {
 		require.Equal(4, len(hashes))
 	})
 	t.Run("empty log", func(t *testing.T) {
-		testData := &filterObject{FromBlock: "1", ToBlock: "4", Address: []string{"0x8ce313ab12bf7aed8136ab36c623ff98c8eaad34"}}
+		testData := &api.FilterObjectWrapper{FromBlock: "1", ToBlock: "4", Address: []string{"0x8ce313ab12bf7aed8136ab36c623ff98c8eaad34"}}
 		filter, err := getTopicsAddress(testData.Address, testData.Topics)
 		require.NoError(err)
 		from, err := strconv.ParseUint(testData.FromBlock, 10, 64)
@@ -59,7 +60,7 @@ func TestLogsInRange(t *testing.T) {
 		require.Equal(0, len(hashes))
 	})
 	t.Run("over 5000 pagenation size", func(t *testing.T) {
-		testData := &filterObject{FromBlock: "1", ToBlock: "4"}
+		testData := &api.FilterObjectWrapper{FromBlock: "1", ToBlock: "4"}
 		filter, err := getTopicsAddress(testData.Address, testData.Topics)
 		require.NoError(err)
 		from, err := strconv.ParseUint(testData.FromBlock, 10, 64)
@@ -73,7 +74,7 @@ func TestLogsInRange(t *testing.T) {
 		require.Equal(4, len(hashes))
 	})
 	t.Run("invalid start and end height", func(t *testing.T) {
-		testData := &filterObject{FromBlock: "2", ToBlock: "1"}
+		testData := &api.FilterObjectWrapper{FromBlock: "2", ToBlock: "1"}
 		filter, err := getTopicsAddress(testData.Address, testData.Topics)
 		require.NoError(err)
 		from, err := strconv.ParseUint(testData.FromBlock, 10, 64)
@@ -87,7 +88,7 @@ func TestLogsInRange(t *testing.T) {
 		require.Equal(expectedErr.Error(), err.Error())
 	})
 	t.Run("start block > tip height", func(t *testing.T) {
-		testData := &filterObject{FromBlock: "5", ToBlock: "5"}
+		testData := &api.FilterObjectWrapper{FromBlock: "5", ToBlock: "5"}
 		filter, err := getTopicsAddress(testData.Address, testData.Topics)
 		require.NoError(err)
 		from, err := strconv.ParseUint(testData.FromBlock, 10, 64)
@@ -110,7 +111,7 @@ func BenchmarkLogsInRange(b *testing.B) {
 	defer ctrl.Finish()
 	blk := mock_blockindex.NewMockBloomFilterIndexer(ctrl)
 
-	testData := &filterObject{FromBlock: "0x1"}
+	testData := &api.FilterObjectWrapper{FromBlock: "0x1"}
 	filter, _ := getTopicsAddress(testData.Address, testData.Topics)
 	from, _ := strconv.ParseInt(testData.FromBlock, 10, 64)
 	to, _ := strconv.ParseInt(testData.ToBlock, 10, 64)
@@ -126,7 +127,7 @@ func BenchmarkLogsInRange(b *testing.B) {
 func getTopicsAddress(addr []string, topics [][]string) (iotexapi.LogsFilter, error) {
 	var filter iotexapi.LogsFilter
 	for _, ethAddr := range addr {
-		ioAddr, err := ethAddrToIoAddr(ethAddr)
+		ioAddr, err := api.EthAddrToIoAddrWrapper(ethAddr)
 		if err != nil {
 			return iotexapi.LogsFilter{}, err
 		}
@@ -135,7 +136,7 @@ func getTopicsAddress(addr []string, topics [][]string) (iotexapi.LogsFilter, er
 	for _, tp := range topics {
 		var topic [][]byte
 		for _, str := range tp {
-			b, err := hexToBytes(str)
+			b, err := api.HexToBytesWrapper(str)
 			if err != nil {
 				return iotexapi.LogsFilter{}, err
 			}
@@ -147,7 +148,7 @@ func getTopicsAddress(addr []string, topics [][]string) (iotexapi.LogsFilter, er
 	return filter, nil
 }
 
-func setupTestCoreSerivce() (CoreService, blockchain.Blockchain, blockdao.BlockDAO, actpool.ActPool, func()) {
+func setupTestCoreSerivce() (api.CoreService, blockchain.Blockchain, blockdao.BlockDAO, actpool.ActPool, func()) {
 	cfg := newConfig()
 
 	// TODO (zhi): revise
@@ -167,10 +168,10 @@ func setupTestCoreSerivce() (CoreService, blockchain.Blockchain, blockdao.BlockD
 		panic(err)
 	}
 
-	opts := []Option{WithBroadcastOutbound(func(ctx context.Context, chainID uint32, msg proto.Message) error {
+	opts := []api.Option{api.WithBroadcastOutbound(func(ctx context.Context, chainID uint32, msg proto.Message) error {
 		return nil
 	})}
-	svr, err := newCoreService(cfg.API, bc, nil, sf, dao, indexer, bfIndexer, ap, registry, opts...)
+	svr, err := api.NewCoreServiceWrapper(cfg.API, bc, nil, sf, dao, indexer, bfIndexer, ap, registry, opts...)
 	if err != nil {
 		panic(err)
 	}
