+++ test/apitest/web3server_integrity_test.go
@@ -3,7 +3,7 @@
 // or fitness for purpose and, to the extent permitted by law, all liability for your use of the code is disclaimed.
 // This source code is governed by Apache License 2.0 that can be found in the LICENSE file.
 
-package api
+package apitest
 
 import (
 	"context"
@@ -18,6 +18,7 @@ import (
 	"github.com/stretchr/testify/require"
 
 	"github.com/iotexproject/iotex-core/actpool"
+	"github.com/iotexproject/iotex-core/api"
 	"github.com/iotexproject/iotex-core/blockchain"
 	"github.com/iotexproject/iotex-core/blockchain/blockdao"
 	"github.com/iotexproject/iotex-core/test/identityset"
@@ -30,12 +31,12 @@ const (
 
 func TestWeb3ServerIntegrity(t *testing.T) {
 	svr, bc, dao, actPool, cleanIndexFile := setupTestServer()
-	web3svr := svr.httpSvr
+	web3svr := api.ServerV2Wrapper(*svr).HttpSvr()
 	defer cleanIndexFile()
 	ctx := context.Background()
 	web3svr.Start(ctx)
 	defer web3svr.Stop(ctx)
-	handler := newHTTPHandler(NewWeb3Handler(svr.core, ""))
+	handler := api.NewHTTPHandlerWrapper(api.NewWeb3Handler(api.ServerV2Wrapper(*svr).Core(), ""))
 
 	// send request
 	t.Run("eth_gasPrice", func(t *testing.T) {
@@ -147,7 +148,7 @@ func TestWeb3ServerIntegrity(t *testing.T) {
 	})
 }
 
-func setupTestServer() (*ServerV2, blockchain.Blockchain, blockdao.BlockDAO, actpool.ActPool, func()) {
+func setupTestServer() (*api.ServerV2, blockchain.Blockchain, blockdao.BlockDAO, actpool.ActPool, func()) {
 	cfg := newConfig()
 	cfg.Chain.EVMNetworkID = _evmNetworkID
 	cfg.API.HTTPPort = testutil.RandomPort()
@@ -157,7 +158,34 @@ func setupTestServer() (*ServerV2, blockchain.Blockchain, blockdao.BlockDAO, act
 	}
 }
 
-func serveTestHTTP(require *require.Assertions, handler *hTTPHandler, method string, param string) interface{} {
+func getStorageAt(t *testing.T, handler *api.HTTPHandlerWrapper, bc blockchain.Blockchain, dao blockdao.BlockDAO, actPool actpool.ActPool) {
+	require := require.New(t)
+	// deploy a contract
+	contractCode := "608060405234801561001057600080fd5b50610150806100206000396000f3fe608060405234801561001057600080fd5b50600436106100365760003560e01c806360fe47b11461003b5780636d4ce63c14610057575b600080fd5b6100556004803603810190610050919061009d565b610075565b005b61005f61007f565b60405161006c91906100d9565b60405180910390f35b8060008190555050565b60008054905090565b60008135905061009781610103565b92915050565b6000602082840312156100b3576100b26100fe565b5b60006100c184828501610088565b91505092915050565b6100d3816100f4565b82525050565b60006020820190506100ee60008301846100ca565b92915050565b6000819050919050565b600080fd5b61010c816100f4565b811461011757600080fd5b5056fea2646970667358221220c86a8c4dd175f55f5732b75b721d714ceb38a835b87c6cf37cf28c790813e19064736f6c63430008070033"
+	contract, _ := deployContractV2(bc, dao, actPool, identityset.PrivateKey(13), 1, bc.TipHeight(), contractCode)
+	contractAddr, _ := api.IoAddrToEthAddrWrapper(contract)
+
+	for _, test := range []struct {
+		params   string
+		expected int
+	}{
+		{fmt.Sprintf(`["%s", "0x0"]`, contractAddr), 1},
+		{`[1]`, 0},
+		{`["TEST", "TEST"]`, 0},
+	} {
+		result := serveTestHTTP(require, handler, "eth_getStorageAt", test.params)
+		if test.expected == 0 {
+			require.Nil(result)
+			return
+		}
+		actual, ok := result.(string)
+		require.True(ok)
+		// the value of any contract at pos0 is be "0x0000000000000000000000000000000000000000000000000000000000000000"
+		require.Equal("0x0000000000000000000000000000000000000000000000000000000000000000", actual)
+	}
+}
+
+func serveTestHTTP(require *require.Assertions, handler *api.HTTPHandlerWrapper, method string, param string) interface{} {
 	req, _ := http.NewRequest(http.MethodPost, "http://url.com",
 		strings.NewReader(fmt.Sprintf(`[{"jsonrpc":"2.0","method":"%s","params":%s,"id":1}]`, method, param)))
 	resp := httptest.NewRecorder()
@@ -175,32 +203,32 @@ func serveTestHTTP(require *require.Assertions, handler *hTTPHandler, method str
 	return vals[0].Result
 }
 
-func gasPrice(t *testing.T, handler *hTTPHandler) {
+func gasPrice(t *testing.T, handler *api.HTTPHandlerWrapper) {
 	require := require.New(t)
 	result := serveTestHTTP(require, handler, "eth_gasPrice", "[]")
 	actual, ok := result.(string)
 	require.True(ok)
-	require.Equal(uint64ToHex(1000000000000), actual)
+	require.Equal(api.Uint64ToHexWrapper(1000000000000), actual)
 }
 
-func chainID(t *testing.T, handler *hTTPHandler) {
+func chainID(t *testing.T, handler *api.HTTPHandlerWrapper) {
 	require := require.New(t)
 	result := serveTestHTTP(require, handler, "eth_chainId", "[]")
 	actual, ok := result.(string)
 	require.True(ok)
-	require.Equal(uint64ToHex(1), actual)
-	require.Equal(uint64ToHex(uint64(_evmNetworkID)), actual)
+	require.Equal(api.Uint64ToHexWrapper(1), actual)
+	require.Equal(api.Uint64ToHexWrapper(uint64(_evmNetworkID)), actual)
 }
 
-func blockNumber(t *testing.T, handler *hTTPHandler) {
+func blockNumber(t *testing.T, handler *api.HTTPHandlerWrapper) {
 	require := require.New(t)
 	result := serveTestHTTP(require, handler, "eth_blockNumber", "[]")
 	actual, ok := result.(string)
 	require.True(ok)
-	require.Equal(uint64ToHex(4), actual)
+	require.Equal(api.Uint64ToHexWrapper(4), actual)
 }
 
-func getBlockByNumber(t *testing.T, handler *hTTPHandler) {
+func getBlockByNumber(t *testing.T, handler *api.HTTPHandlerWrapper) {
 	require := require.New(t)
 	for _, test := range []struct {
 		params   string
@@ -221,7 +249,7 @@ func getBlockByNumber(t *testing.T, handler *hTTPHandler) {
 	}
 }
 
-func getBalance(t *testing.T, handler *hTTPHandler) {
+func getBalance(t *testing.T, handler *api.HTTPHandlerWrapper) {
 	require := require.New(t)
 	result := serveTestHTTP(require, handler, "eth_getBalance", `["0xDa7e12Ef57c236a06117c5e0d04a228e7181CF36", 1]`)
 	ans, ok := new(big.Int).SetString("9999999999999999999999999991", 10)
@@ -233,7 +261,7 @@ func getBalance(t *testing.T, handler *hTTPHandler) {
 	require.Equal("0x"+fmt.Sprintf("%x", ans), actual)
 }
 
-func getTransactionCount(t *testing.T, handler *hTTPHandler) {
+func getTransactionCount(t *testing.T, handler *api.HTTPHandlerWrapper) {
 	require := require.New(t)
 	for _, test := range []struct {
 		params   string
@@ -245,11 +273,11 @@ func getTransactionCount(t *testing.T, handler *hTTPHandler) {
 		result := serveTestHTTP(require, handler, "eth_getTransactionCount", test.params)
 		actual, ok := result.(string)
 		require.True(ok)
-		require.Equal(uint64ToHex(uint64(test.expected)), actual)
+		require.Equal(api.Uint64ToHexWrapper(uint64(test.expected)), actual)
 	}
 }
 
-func ethCall(t *testing.T, handler *hTTPHandler) {
+func ethCall(t *testing.T, handler *api.HTTPHandlerWrapper) {
 	require := require.New(t)
 	for _, test := range []struct {
 		params   string
@@ -291,7 +319,7 @@ func ethCall(t *testing.T, handler *hTTPHandler) {
 	}
 }
 
-func getNodeInfo(t *testing.T, handler *hTTPHandler) {
+func getNodeInfo(t *testing.T, handler *api.HTTPHandlerWrapper) {
 	require := require.New(t)
 	result := serveTestHTTP(require, handler, "web3_clientVersion", "[]")
 	actual, ok := result.(string)
@@ -299,83 +327,56 @@ func getNodeInfo(t *testing.T, handler *hTTPHandler) {
 	require.Equal("NoBuildInfo/NoBuildInfo", actual)
 }
 
-func getBlockTransactionCountByHash(t *testing.T, handler *hTTPHandler, bc blockchain.Blockchain) {
+func getBlockTransactionCountByHash(t *testing.T, handler *api.HTTPHandlerWrapper, bc blockchain.Blockchain) {
 }
 
-func getBlockByHash(t *testing.T, handler *hTTPHandler, bc blockchain.Blockchain) {
+func getBlockByHash(t *testing.T, handler *api.HTTPHandlerWrapper, bc blockchain.Blockchain) {
 }
 
-func getTransactionByHash(t *testing.T, handler *hTTPHandler) {
+func getTransactionByHash(t *testing.T, handler *api.HTTPHandlerWrapper) {
 }
 
-func getLogs(t *testing.T, handler *hTTPHandler) {
+func getLogs(t *testing.T, handler *api.HTTPHandlerWrapper) {
 }
 
-func getTransactionReceipt(t *testing.T, handler *hTTPHandler) {
+func getTransactionReceipt(t *testing.T, handler *api.HTTPHandlerWrapper) {
 }
 
-func getBlockTransactionCountByNumber(t *testing.T, handler *hTTPHandler) {
+func getBlockTransactionCountByNumber(t *testing.T, handler *api.HTTPHandlerWrapper) {
 }
 
-func getTransactionByBlockHashAndIndex(t *testing.T, handler *hTTPHandler, bc blockchain.Blockchain) {
+func getTransactionByBlockHashAndIndex(t *testing.T, handler *api.HTTPHandlerWrapper, bc blockchain.Blockchain) {
 }
 
-func getTransactionByBlockNumberAndIndex(t *testing.T, handler *hTTPHandler) {
+func getTransactionByBlockNumberAndIndex(t *testing.T, handler *api.HTTPHandlerWrapper) {
 }
 
-func newfilter(t *testing.T, handler *hTTPHandler) {
+func newfilter(t *testing.T, handler *api.HTTPHandlerWrapper) {
 }
 
-func newBlockFilter(t *testing.T, handler *hTTPHandler) {
+func newBlockFilter(t *testing.T, handler *api.HTTPHandlerWrapper) {
 }
 
-func getFilterChanges(t *testing.T, handler *hTTPHandler) {
+func getFilterChanges(t *testing.T, handler *api.HTTPHandlerWrapper) {
 }
 
-func getFilterLogs(t *testing.T, handler *hTTPHandler) {
+func getFilterLogs(t *testing.T, handler *api.HTTPHandlerWrapper) {
 }
 
-func getNetworkID(t *testing.T, handler *hTTPHandler) {
+func getNetworkID(t *testing.T, handler *api.HTTPHandlerWrapper) {
 }
 
-func ethAccounts(t *testing.T, handler *hTTPHandler) {
+func ethAccounts(t *testing.T, handler *api.HTTPHandlerWrapper) {
 }
 
-func web3Staking(t *testing.T, handler *hTTPHandler) {
+func web3Staking(t *testing.T, handler *api.HTTPHandlerWrapper) {
 }
 
-func sendRawTransaction(t *testing.T, handler *hTTPHandler) {
+func sendRawTransaction(t *testing.T, handler *api.HTTPHandlerWrapper) {
 }
 
-func estimateGas(t *testing.T, handler *hTTPHandler, bc blockchain.Blockchain, dao blockdao.BlockDAO, actPool actpool.ActPool) {
+func estimateGas(t *testing.T, handler *api.HTTPHandlerWrapper, bc blockchain.Blockchain, dao blockdao.BlockDAO, actPool actpool.ActPool) {
 }
 
-func getCode(t *testing.T, handler *hTTPHandler, bc blockchain.Blockchain, dao blockdao.BlockDAO, actPool actpool.ActPool) {
-}
-
-func getStorageAt(t *testing.T, handler *hTTPHandler, bc blockchain.Blockchain, dao blockdao.BlockDAO, actPool actpool.ActPool) {
-	require := require.New(t)
-	// deploy a contract
-	contractCode := "608060405234801561001057600080fd5b50610150806100206000396000f3fe608060405234801561001057600080fd5b50600436106100365760003560e01c806360fe47b11461003b5780636d4ce63c14610057575b600080fd5b6100556004803603810190610050919061009d565b610075565b005b61005f61007f565b60405161006c91906100d9565b60405180910390f35b8060008190555050565b60008054905090565b60008135905061009781610103565b92915050565b6000602082840312156100b3576100b26100fe565b5b60006100c184828501610088565b91505092915050565b6100d3816100f4565b82525050565b60006020820190506100ee60008301846100ca565b92915050565b6000819050919050565b600080fd5b61010c816100f4565b811461011757600080fd5b5056fea2646970667358221220c86a8c4dd175f55f5732b75b721d714ceb38a835b87c6cf37cf28c790813e19064736f6c63430008070033"
-	contract, _ := deployContractV2(bc, dao, actPool, identityset.PrivateKey(13), 1, bc.TipHeight(), contractCode)
-	contractAddr, _ := ioAddrToEthAddr(contract)
-
-	for _, test := range []struct {
-		params   string
-		expected int
-	}{
-		{fmt.Sprintf(`["%s", "0x0"]`, contractAddr), 1},
-		{`[1]`, 0},
-		{`["TEST", "TEST"]`, 0},
-	} {
-		result := serveTestHTTP(require, handler, "eth_getStorageAt", test.params)
-		if test.expected == 0 {
-			require.Nil(result)
-			return
-		}
-		actual, ok := result.(string)
-		require.True(ok)
-		// the value of any contract at pos0 is be "0x0000000000000000000000000000000000000000000000000000000000000000"
-		require.Equal("0x0000000000000000000000000000000000000000000000000000000000000000", actual)
-	}
+func getCode(t *testing.T, handler *api.HTTPHandlerWrapper, bc blockchain.Blockchain, dao blockdao.BlockDAO, actPool actpool.ActPool) {
 }
